pyproject.toml
[tool.poetry]
name = "dom-kyrs-7"
version = "0.1.0"
description = "Домашка по модулю курса №7"
authors = ["Qiaxx"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.4.2"
ipython = "^8.26.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
django = "^5.0.6"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.7"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-filter = "^24.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
